accessible/tests/mochitest/states/test_aria_imgmap.html

<!DOCTYPE html>
<html>
<head>
  <title>Test usemap elements and ARIA</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../states.js"></script>

  <script type="application/javascript">
  function doTest()
  {
    ensureImageMapTree("imagemap");

    var imageMap = getAccessible("imagemap");

    var t1 = imageMap.getChildAt(0);
    testStates(t1, 0, EXT_STATE_EDITABLE, STATE_LINKED);
    var t2 = imageMap.getChildAt(1);
    testStates(t2, 0, EXT_STATE_EDITABLE, STATE_LINKED);
    var rb1 = imageMap.getChildAt(2);
    testStates(rb1, (STATE_CHECKABLE | STATE_CHECKED), 0, STATE_LINKED);
    var rb2 = imageMap.getChildAt(3);
    testStates(rb2, STATE_CHECKABLE, 0, STATE_CHECKED, STATE_LINKED);
    var cb1 = imageMap.getChildAt(4);
    testStates(cb1, (STATE_CHECKABLE | STATE_CHECKED), 0, STATE_LINKED);
    var cbox = imageMap.getChildAt(5);
    testStates(cbox, (STATE_HASPOPUP | STATE_COLLAPSED),
               EXT_STATE_EXPANDABLE, STATE_LINKED);

    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
  addA11yLoadEvent(doTest);
  </script>
</head>
<body>

<a target="_blank"
  href="https://bugzilla.mozilla.org/show_bug.cgi?id=548291"
  title="ARIA states on image maps">
Mozilla Bug 548291
</a>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
</pre>

<img id="imagemap" src="../formimage.png" width="219" height="229" border="0" usemap="#ariaMap">
<map id="ariaMap" name="ariaMap">
  <area id="t1" role="textbox" shape="rect" tabindex="0" alt="" title="first name"  coords="4,20,108,48" href="#" />
  <area id="t2" role="textbox" shape="rect" alt="" title="last name"  coords="111,21,215,50" href="#" />
  <area id="rb1" role="radio" aria-checked="true" shape="circle" alt="" title="male"  coords="60,75,11" href="#" />
  <area id="rb2" role="radio" shape="circle" alt="" title="female"  coords="73,94,11" href="#" />
  <area id="cb1" role="checkbox" aria-checked="true" shape="rect" alt="" title="have bike"  coords="95,123,118,145" href="#" />
  <area id="cbox" role="combobox" shape="rect" alt="" title="bike model"  coords="120,124,184,146" href="#" />
  <area id="cb2" role="checkbox" shape="rect" alt="" title="have car"  coords="90,145,114,164" href="#" />
  <area id="cb3" role="checkbox" shape="rect" alt="" title="have airplane"  coords="130,163,152,184" href="#" />
  <area id="b1" role="button" shape="rect" alt="" title="submit"  coords="4,198,67,224" href="#" />
</map>
</body>
</html>
